Parse .arktribe and .arkprofile data from .ark save when using --UseStore.

+++ SavegameToolkit/GameObject.cs
@@ -279,7 +279,7 @@ namespace SavegameToolkit {
             archive.HasUnknownData = true;
         }
 
-        public void LoadProperties(ArkArchive archive, GameObject next, int propertiesBlockOffset) {
+        public void LoadProperties(ArkArchive archive, GameObject next, long propertiesBlockOffset) {
             long offset = propertiesBlockOffset + propertiesOffset;
             long nextOffset = propertiesBlockOffset + next?.propertiesOffset ?? archive.Limit;
